requirements.txt
# Flask and Core Dependencies
Flask==3.0.0
flask-cors==4.0.0

# Database
supabase==2.0.0
psycopg2-binary==2.9.9

# File Processing
pandas==2.1.4
openpyxl==3.1.2
pdf-parse==1.3.11
python-docx==1.1.0
PyPDF2==3.0.1  # Alternative PDF parser
pdfplumber==0.10.3  # Another PDF parser option

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

